--- src/componenets/forms/Form.css
@@ -1,10 +1,0 @@
-.header-new-task {
-    text-align: Center;
-  color: #FE5401;
-  font-size: 30px;
-}
-.header-new-event {
-    text-align: Center;
-  color: #FE5401;
-  font-size: 30px;
-}
