+++ src/components/forms/Form.css
@@ -1,0 +1,10 @@
+.header-new-task {
+    text-align: Center;
+  color: #FE5401;
+  font-size: 30px;
+}
+.header-new-event {
+    text-align: Center;
+  color: #FE5401;
+  font-size: 30px;
+}
